refactor torch model into seperate file
